requirements.txt
Django==4.2
djangorestframework==3.15.1
django-ninja==1.1.0